Create account bg and separate wallet bg logic
