+++ src/background/walletBackground.ts
@@ -1,18 +1,14 @@
 import scrypt from 'scryptsy';
-import { Wallet, Insight, Network } from 'qtumjs-wallet';
-import { isEmpty, split, find } from 'lodash';
+import { Wallet, Insight } from 'qtumjs-wallet';
+import { isEmpty, split } from 'lodash';
 import axios from 'axios';
 
 import Background from '.';
 import { MESSAGE_TYPE, STORAGE } from '../constants';
-import Account from '../models/Account';
 
 const INIT_VALUES = {
   appSalt: undefined,
   passwordHash: undefined,
-  mainnetAccounts: [],
-  testnetAccounts: [],
-  loggedInAccount: undefined,
   wallet: undefined,
   info: undefined,
 };
@@ -23,15 +19,8 @@ export default class WalletBackground {
   private static GET_INFO_INTERVAL_MS: number = 10000;
   private static GET_PRICE_INTERVAL_MS: number = 60000;
 
-  public loggedInAccount?: Account = INIT_VALUES.loggedInAccount;
   public wallet?: Wallet = INIT_VALUES.wallet;
   public info?: Insight.IGetInfo = INIT_VALUES.info;
-  public get accounts(): Account[] {
-    return this.bg.network.isMainNet ? this.mainnetAccounts : this.testnetAccounts;
-  }
-  public get hasAccounts(): boolean {
-    return !isEmpty(this.mainnetAccounts) || !isEmpty(this.testnetAccounts);
-  }
   public get qtumBalanceUSD(): string {
     if (this.qtumPriceUSD && this.info) {
       return (this.qtumPriceUSD * this.info.balance).toFixed(2);
@@ -43,8 +32,6 @@ export default class WalletBackground {
   private bg: Background;
   private appSalt?: Uint8Array = INIT_VALUES.appSalt;
   private passwordHash?: string = INIT_VALUES.passwordHash;
-  private mainnetAccounts: Account[] = INIT_VALUES.mainnetAccounts;
-  private testnetAccounts: Account[] = INIT_VALUES.testnetAccounts;
   private getInfoInterval?: number = undefined;
   private getPriceInterval?: number = undefined;
   private qtumPriceUSD: number = 0;
@@ -60,161 +47,31 @@ export default class WalletBackground {
     chrome.runtime.onMessage.addListener(this.handleMessage);
 
     // Initializes all the values from Chrome storage on startup
-    const { APP_SALT, MAINNET_ACCOUNTS, TESTNET_ACCOUNTS } = STORAGE;
-    chrome.storage.local.get([APP_SALT, MAINNET_ACCOUNTS, TESTNET_ACCOUNTS],
-      ({ appSalt, mainnetAccounts, testnetAccounts }: any) => {
-        if (!isEmpty(appSalt)) {
-          const array = split(appSalt, ',').map((str) => parseInt(str, 10));
-          this.appSalt =  Uint8Array.from(array);
-        }
+    const { APP_SALT } = STORAGE;
+    chrome.storage.local.get([APP_SALT], ({ appSalt }: any) => {
+      if (!isEmpty(appSalt)) {
+        const array = split(appSalt, ',').map((str) => parseInt(str, 10));
+        this.appSalt =  Uint8Array.from(array);
+      }
 
-        if (!isEmpty(mainnetAccounts)) {
-          this.mainnetAccounts = mainnetAccounts;
-        }
-
-        if (!isEmpty(testnetAccounts)) {
-          this.testnetAccounts = testnetAccounts;
-        }
-
-        // Show the Login page after fetching storage
-        chrome.runtime.sendMessage({ type: MESSAGE_TYPE.ROUTE_LOGIN });
-      });
-  }
-
-  public isWalletNameTaken = (name: string): boolean => {
-    return !!find(this.accounts, { name });
-  }
-
-  public login = async (password: string) => {
-    this.generateAppSaltIfNecessary();
-
-    try {
-      await this.derivePasswordHash(password);
-    } catch (err) {
-      throw err;
-    }
-
-    if (!this.hasAccounts) {
-      // New user. No created wallets yet. No need to validate.
-      this.routeToAccountPage();
-      return;
-    }
-
-    const isPwValid = await this.validatePassword();
-    if (isPwValid) {
-      this.routeToAccountPage();
-      return;
-    }
-
-    chrome.runtime.sendMessage({ type: MESSAGE_TYPE.LOGIN_FAILURE });
-  }
-
-  /*
-  * Creates an account, stores it, and logs in.
-  * @param accountName {string} The account name for the new wallet account.
-  * @param mnemonic {string} The mnemonic to derive the wallet from.
-  */
-  public async addAccountAndLogin(accountName: string, mnemonic: string) {
-    // Get encrypted private key
-    const network = this.bg.network.network;
-    this.wallet = await network.fromMnemonic(mnemonic);
-    const privateKeyHash = await this.wallet.toEncryptedPrivateKey(
-      this.validPasswordHash,
-      WalletBackground.SCRYPT_PARAMS_PRIV_KEY,
-    );
-    const account = new Account(accountName, privateKeyHash);
-
-    // Add account if not existing
-    if (this.bg.network.isMainNet) {
-      this.mainnetAccounts.push(account);
-      chrome.storage.local.set({
-        [STORAGE.MAINNET_ACCOUNTS]: this.mainnetAccounts,
-      }, () => console.log('Mainnet Account added', account));
-    } else {
-      this.testnetAccounts.push(account);
-      chrome.storage.local.set({
-        [STORAGE.TESTNET_ACCOUNTS]: this.testnetAccounts,
-      }, () => console.log('Testnet Account added', account));
-    }
-
-    this.loggedInAccount = account;
-    await this.onAccountLoggedIn();
-  }
-
-  public importMnemonic = async (accountName: string, mnemonic: string) => {
-    const isTaken = await this.isWalletMnemonicTaken(mnemonic);
-    if (isTaken) {
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.IMPORT_MNEMONIC_FAILURE });
-      return;
-    }
-
-    this.addAccountAndLogin(accountName, mnemonic);
-  }
-
-  public saveToFile = (accountName: string, mnemonic: string) => {
-    const timestamp = new Date().toLocaleDateString(undefined, {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-      second: '2-digit',
-      hour12: false,
+      // TODO: wait for all init before routing to login. move this to upper level Background
+      // Show the Login page after fetching storage
+      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.ROUTE_LOGIN });
     });
-    const file = new Blob([mnemonic], {type: 'text/plain'});
-    const element = document.createElement('a');
-    element.href = URL.createObjectURL(file);
-    element.download = `qrypto_${accountName}_${timestamp}.bak`;
-    element.click();
-
-    this.addAccountAndLogin(accountName, mnemonic);
   }
 
   /*
-  * Finds the account based on the name and logs in.
-  * @param accountName {string} The account name to search by.
+  * Resets the wallet vars back to the initial state.
   */
-  public async loginAccount(accountName: string) {
-    const accounts = this.bg.network.isMainNet ? this.mainnetAccounts : this.testnetAccounts;
-    const foundAccount = find(accounts, { name: accountName });
-
-    if (!foundAccount) {
-      throw Error('Account should not be undefined');
-    }
-
-    this.loggedInAccount = foundAccount;
-
-    // Recover wallet
-    const network = this.bg.network.network;
-    this.wallet = await network.fromEncryptedPrivateKey(
-      this.loggedInAccount!.privateKeyHash,
-      this.validPasswordHash,
-      WalletBackground.SCRYPT_PARAMS_PRIV_KEY,
-    );
-
-    await this.onAccountLoggedIn();
-  }
-
-  public logout = () => {
-    this.stopPolling();
-    this.loggedInAccount = INIT_VALUES.loggedInAccount;
+  public resetWallet = () => {
     this.wallet = INIT_VALUES.wallet;
     this.info =  INIT_VALUES.info;
-    this.routeToAccountPage();
   }
 
-  public sendTokens = async (receiverAddress: string, amount: number) => {
-    try {
-      const wallet = this.wallet!;
-      await wallet.send(receiverAddress, amount * 1e8, { feeRate: 4000 });
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.SEND_TOKENS_SUCCESS });
-    } catch (err) {
-      console.log(err);
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.SEND_TOKENS_FAILURE, error: err });
-    }
-  }
-
-  private generateAppSaltIfNecessary = () => {
+  /*
+  * Generates the one-time created appSalt (if necessary) used to encrypt the user password.
+  */
+  public generateAppSaltIfNecessary = () => {
     try {
       if (!this.appSalt) {
         const appSalt: Uint8Array = window.crypto.getRandomValues(new Uint8Array(16)) as Uint8Array;
@@ -231,9 +88,9 @@ export default class WalletBackground {
 
   /*
   * Derives the password hash with the password input.
-  * @return Promise undefined or error.
+  * @return Undefined or error.
   */
-  private derivePasswordHash = async (password: string): Promise<any> => {
+  public derivePasswordHash = async (password: string): Promise<any> => {
     return new Promise((resolve: any, reject: any) => {
       setTimeout(() => {
         try {
@@ -255,74 +112,40 @@ export default class WalletBackground {
   }
 
   /*
-  * Routes to the CreateWallet or AccountLogin page after unlocking with the password.
+  * Derives the private key hash with the password hash.
+  * @return Private key hash or exception thrown.
   */
-  private routeToAccountPage = () => {
-    const accounts = this.bg.network.isMainNet ? this.mainnetAccounts : this.testnetAccounts;
-    if (isEmpty(accounts)) {
-      // Account not found, route to Create Wallet page
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.LOGIN_SUCCESS_NO_ACCOUNTS });
-    } else {
-      // Accounts found, route to Account Login page
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.LOGIN_SUCCESS_WITH_ACCOUNTS });
+  public derivePrivateKeyHash = async (mnemonic: string): Promise<string> => {
+    try {
+      const network = this.bg.network.network;
+      this.wallet = await network.fromMnemonic(mnemonic);
+      const privateKeyHash = await this.wallet.toEncryptedPrivateKey(
+        this.validPasswordHash,
+        WalletBackground.SCRYPT_PARAMS_PRIV_KEY,
+      );
+      return privateKeyHash;
+    } catch (err) {
+      throw err;
     }
   }
 
   /*
-  * Validates a password by decrypting a private key hash into a wallet instance.
-  * @return Is the password valid.
+  * Recovers the wallet instance from an encrypted private key.
+  * @param privateKeyHash The private key hash to recover the wallet from.
   */
-  private validatePassword = async (): Promise<boolean> => {
-    let account: Account;
-    if (!isEmpty(this.mainnetAccounts)) {
-      account = this.mainnetAccounts[0];
-    } else if (!isEmpty(this.testnetAccounts)) {
-      account = this.testnetAccounts[0];
-    } else {
-      throw Error('Trying to validate password without existing account');
-    }
-
-    try {
-      const network: Network = this.bg.network.network;
-      await network.fromEncryptedPrivateKey(
-        account.privateKeyHash,
-        this.validPasswordHash,
-        WalletBackground.SCRYPT_PARAMS_PRIV_KEY,
-      );
-      return true;
-    } catch (err) {
-      console.log(err);
-      return false;
-    }
-  }
-
-  private isWalletMnemonicTaken = async (mnemonic: string): Promise<boolean> => {
+  public async recoverWallet(privateKeyHash: string) {
     const network = this.bg.network.network;
-    const wallet = await network.fromMnemonic(mnemonic);
-    const privateKeyHash = await wallet.toEncryptedPrivateKey(
+    this.wallet = await network.fromEncryptedPrivateKey(
+      privateKeyHash,
       this.validPasswordHash,
       WalletBackground.SCRYPT_PARAMS_PRIV_KEY,
     );
-    const accounts = this.bg.network.isMainNet ? this.mainnetAccounts : this.testnetAccounts;
-    return !!find(accounts, { privateKeyHash });
   }
 
-  private getQtumPrice = async () => {
-    try {
-      const jsonObj = await axios.get('https://api.coinmarketcap.com/v2/ticker/1684/');
-      this.qtumPriceUSD = jsonObj.data.data.quotes.USD.price;
-      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.GET_QTUM_PRICE_RETURN, qtumBalanceUSD: this.qtumBalanceUSD });
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
-  private getWalletInfo = async () => {
-    this.info = await this.wallet!.getInfo();
-    chrome.runtime.sendMessage({ type: MESSAGE_TYPE.GET_WALLET_INFO_RETURN, info: this.info });
-  }
-
-  private startPolling = async () => {
+  /*
+  * Starts polling for periodic info updates.
+  */
+  public startPolling = async () => {
     await this.getWalletInfo();
     await this.getQtumPrice();
 
@@ -334,7 +157,10 @@ export default class WalletBackground {
     }, WalletBackground.GET_PRICE_INTERVAL_MS);
   }
 
-  private stopPolling = () => {
+  /*
+  * Stops polling for the periodic info updates.
+  */
+  public stopPolling = () => {
     if (this.getInfoInterval) {
       clearInterval(this.getInfoInterval);
       this.getInfoInterval = undefined;
@@ -346,53 +172,52 @@ export default class WalletBackground {
   }
 
   /*
-  * Actions after adding a new account or logging into an existing account.
+  * Executes a sendtoaddress.
+  * @param receiverAddress The address to send Qtum to.
+  * @param amount The amount to send.
   */
-  private onAccountLoggedIn = async () => {
-    await this.startPolling();
-    chrome.runtime.sendMessage({ type: MESSAGE_TYPE.ACCOUNT_LOGIN_SUCCESS });
+  public sendTokens = async (receiverAddress: string, amount: number) => {
+    try {
+      const wallet = this.wallet!;
+      await wallet.send(receiverAddress, amount * 1e8, { feeRate: 4000 });
+      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.SEND_TOKENS_SUCCESS });
+    } catch (err) {
+      console.log(err);
+      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.SEND_TOKENS_FAILURE, error: err });
+    }
+  }
+
+  /*
+  * Fetches the wallet info from the current wallet instance.
+  */
+  private getWalletInfo = async () => {
+    this.info = await this.wallet!.getInfo();
+    chrome.runtime.sendMessage({ type: MESSAGE_TYPE.GET_WALLET_INFO_RETURN, info: this.info });
+  }
+
+  /*
+  * Gets the current Qtum market price.
+  */
+  private getQtumPrice = async () => {
+    try {
+      const jsonObj = await axios.get('https://api.coinmarketcap.com/v2/ticker/1684/');
+      this.qtumPriceUSD = jsonObj.data.data.quotes.USD.price;
+      chrome.runtime.sendMessage({ type: MESSAGE_TYPE.GET_QTUM_PRICE_RETURN, qtumBalanceUSD: this.qtumBalanceUSD });
+    } catch (err) {
+      console.log(err);
+    }
   }
 
   private handleMessage = (request: any, _: chrome.runtime.MessageSender, sendResponse: (response: any) => void) => {
     switch (request.type) {
-      case MESSAGE_TYPE.LOGIN:
-        this.login(request.password);
-        break;
-      case MESSAGE_TYPE.IMPORT_MNEMONIC:
-        this.importMnemonic(request.accountName, request.mnemonic);
-        break;
-      case MESSAGE_TYPE.CREATE_WALLET:
-        this.addAccountAndLogin(request.accountName, request.mnemonic);
-        break;
-      case MESSAGE_TYPE.SAVE_TO_FILE:
-        this.saveToFile(request.accountName, request.mnemonic);
-        break;
-      case MESSAGE_TYPE.ACCOUNT_LOGIN:
-        this.loginAccount(request.selectedWalletName);
-        break;
-      case MESSAGE_TYPE.SEND_TOKENS:
-        this.sendTokens(request.receiverAddress, request.amount);
-        break;
-      case MESSAGE_TYPE.LOGOUT:
-        this.logout();
-        break;
-      case MESSAGE_TYPE.HAS_ACCOUNTS:
-        sendResponse(this.hasAccounts);
-        break;
-      case MESSAGE_TYPE.GET_ACCOUNTS:
-        sendResponse(this.accounts);
-        break;
-      case MESSAGE_TYPE.GET_LOGGED_IN_ACCOUNT:
-        sendResponse(this.loggedInAccount);
-        break;
       case MESSAGE_TYPE.GET_WALLET_INFO:
         sendResponse(this.info);
         break;
       case MESSAGE_TYPE.GET_QTUM_BALANCE_USD:
         sendResponse(this.qtumBalanceUSD);
         break;
-      case MESSAGE_TYPE.VALIDATE_WALLET_NAME:
-        sendResponse(this.isWalletNameTaken(request.name));
+      case MESSAGE_TYPE.SEND_TOKENS:
+        this.sendTokens(request.receiverAddress, request.amount);
         break;
       default:
         break;
